+++ apps/vecadd/vecadd-brisbane.cpp
@@ -4,13 +4,13 @@
 #include <malloc.h>
 
 int main(int argc, char** argv) {
-    int SIZE;
+    size_t SIZE;
     int *A, *B, *C, *D, *E;
     int ERROR = 0;
 
     brisbane_init(&argc, &argv);
 
-    SIZE = argc > 1 ? atoi(argv[1]) : 16;
+    SIZE = argc > 1 ? atol(argv[1]) : 16;
     printf("SIZE[%d]\n", SIZE);
 
     A = (int*) valloc(SIZE * sizeof(int));
